requirements.txt
torch
transformers
pillow
opensearch-py
pandas
tqdm
sentence_transformers
datasets
fastapi
uvicorn
aiofiles